Collapse `aggregatestore` into `persistence`.
